divyamrutnaturals/urls.py
from django.contrib import admin
from django.urls import path, re_path, include
from django.conf import settings
from django.conf.urls.static import static
from django.shortcuts import redirect
from rest_framework import permissions
from drf_yasg.views import get_schema_view
from drf_yasg import openapi

# Swagger / Redoc schema
schema_view = get_schema_view(
    openapi.Info(
        title="VIVAH",
        default_version='v1',
        description="API documentation",
    ),
    public=True,
    permission_classes=(permissions.AllowAny,),
)

# Redirect function for unmatched URLs
def redirect_to_home(request, path=None):
    return redirect('home')

urlpatterns = [
    path('admin-987/', admin.site.urls),
    path('', include('accounts.api_urls')),

    # Swagger & Redoc
    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
    path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
    path('swagger.json', schema_view.without_ui(cache_timeout=0), name='schema-json'),
]

# Serve static & media files in DEBUG mode
if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)

# Catch-all pattern for unmatched URLs (put **at the very end**)
urlpatterns += [
    re_path(r'^(?P<path>.*)$', redirect_to_home),
]
