requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.3.1
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-countries==7.6.1
django-filter==24.3
django-js-asset==3.0.1
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
firebase-admin==6.6.0
frozenlist==1.5.0
geonamescache==2.0.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
kombu==5.5.1
msgpack==1.1.0
multidict==6.2.0
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pyfcm==1.4.7
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
rsa==4.9
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
